# src/build/swagger.json
{
	"components": {
...
		"schemas": {
...
			"BaseResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"success": {
						"type": "boolean"
					}
				},
				"required": [
					"message",
					"success"
				],
				"type": "object"
			},
			"UserLoginResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"success": {
						"type": "boolean"
					},
					"token": {
						"type": "string"
					}
				},
				"required": [
					"message",
					"success",
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CommentsResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"success": {
						"type": "boolean"
					},
					"id": {
						"type": "string"
					},
					"text": {
						"type": "string",
						"nullable": true
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"userId": {
						"type": "string"
					},
					"collectionId": {
						"type": "string"
					}
				},
				"required": [
					"message",
					"success",
					"id",
					"text",
					"rating",
					"userId",
					"collectionId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CommentsCreateRequest": {
				"properties": {
					"text": {
						"type": "string"
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"collectionId": {
						"type": "string"
					}
				},
				"required": [
					"text",
					"rating",
					"collectionId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CommentsUpdateRequest": {
				"properties": {
					"id": {
						"type": "string"
					},
					"text": {
						"type": "string"
					},
					"rating": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"text",
					"rating"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CommentsListResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"success": {
						"type": "boolean"
					},
					"comments": {
						"items": {
							"properties": {
								"collectionId": {
									"type": "string"
								},
								"userId": {
									"type": "string"
								},
								"rating": {
									"type": "number",
									"format": "double"
								},
								"text": {
									"type": "string",
									"nullable": true
								},
								"id": {
									"type": "string"
								}
							},
							"required": [
								"collectionId",
								"userId",
								"rating",
								"text",
								"id"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"message",
					"success",
					"comments"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "x-access-token",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "nft-collection",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/comments/create": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/CommentsResponse"
										},
										{
											"$ref": "#/components/schemas/BaseResponse"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"message": "Could not create comment"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserLoginResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"message": "Invalid credentials",
											"token": ""
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"message": "Internal Server Error"
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Comments"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CommentsCreateRequest"
							}
						}
					}
				}
			}
		},
		"/comments/update": {
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/CommentsResponse"
										},
										{
											"$ref": "#/components/schemas/BaseResponse"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"message": "Could not update comment"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserLoginResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"message": "Invalid credentials",
											"token": ""
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"message": "Internal Server Error"
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Comments"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CommentsUpdateRequest"
							}
						}
					}
				}
			}
		},
		"/comments/delete": {
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"message": "Could not delete comment"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserLoginResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"message": "Invalid credentials",
											"token": ""
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"message": "Internal Server Error"
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Comments"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CommentsUpdateRequest"
							}
						}
					}
				}
			}
		},
		"/comments/collection/{collectionId}": {
			"get": {
				"operationId": "GetCommentsByCollectionId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/CommentsListResponse"
										},
										{
											"$ref": "#/components/schemas/BaseResponse"
										}
									]
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"message": "Could not retrieve comments"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"message": "Internal Server Error"
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Comments"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "collectionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/comments/get": {
			"get": {
				"operationId": "GetCommentsByUserId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/CommentsListResponse"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"message": "Could not retrieve comments"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"message": "Internal Server Error"
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Comments"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/comments/public/collection/{collectionId}": {
			"get": {
				"operationId": "GetPublicCommentsByCollectionId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/CommentsListResponse"
										},
										{
											"$ref": "#/components/schemas/BaseResponse"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"message": "Could not retrieve comments"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"message": "Internal Server Error"
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Comments"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "collectionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}